requirements/dev.txt
check-manifest==0.37
coverage==4.5.1
docutils==0.14
pytest-flake8==1.0.4
tox==3.1.2
twine==1.11.0
